fixing wave manager

+++ src/public/wave.h
@@ -1,6 +1,7 @@
 #include <genesis.h>
 #include <sprite_eng.h>
 #include <maths.h>
+#include <tools.h>
 
 #include "public/node.h"
 #include "public/library.h"
@@ -11,7 +12,7 @@
 
 #define MAX_ENEMIES 1
 
-extern int EnemyCount;
+extern u16 EnemyCount;
 extern Character* Enemies[MAX_ENEMIES];
 
 u16 WaveManagerInit(u16 VRAMIndex);
